feat(): added script to version all packages fix(control): fix WhiteSpaceValidator to not validate when empty string
