+++ projects/stlmpp/control/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@stlmpp/control",
-  "version": "7.0.0-next.1",
+  "version": "7.0.0-next.2",
   "peerDependencies": {
     "@angular/common": ">= 12 < 13",
     "@angular/core": ">= 12 < 13"
