+++ projects/stlmpp/control/src/lib/validator/white-space/white-space.ts
@@ -1,5 +1,4 @@
 import { ControlValidator } from '../validator';
-import { isNil } from 'st-utils';
 import { Observable } from 'rxjs';
 import { Control } from '../../control/control';
 import { Nullable } from '../../util';
@@ -7,7 +6,7 @@ import { Nullable } from '../../util';
 export class AbstractWhiteSpaceValidator extends ControlValidator<Nullable<string>, boolean> {
   readonly name = 'whiteSpace';
   validate({ value }: Control<Nullable<string>>): Observable<boolean | null> | boolean | null {
-    if (isNil(value)) {
+    if (!value) {
       return null;
     }
     return value.trim() === '' || null;
